refactors project so only one firestore instance is created
